added ability to calculate average based on propertyname

+++ IOT_DeviceManager.API/Entity/ThermostatDevice/ThermostatDevice.cs
@@ -10,7 +10,7 @@ namespace IOT_DeviceManager.API.Entity.ThermostatDevice
         public string Id { get; set; }
         public string DeviceName { get; set; }
         public IDeviceStatus Status { get; set; }
-        public float SetPoint { get; set; }
+        public double SetPoint { get; set; }
         public IEnumerable<IDeviceMeasurement> Measurements { get; set; } = new List<ThermostatDeviceMeasurement>();
     }
 }
